Allow config tracking toggling - Add enable-config-tracking command arg (true by default) - Add no-operation tracker - Add tests for nop tracker
